считываем из файла settings.txt и заполняем таблицу при открытии

+++ mainwindow.cpp
@@ -2,6 +2,7 @@
 #include "ui_mainwindow.h"
 #include <QLayout>
 #include <QMessageBox>
+
 extern QPixmap iconUp1;
 
 QVBoxLayout *vBoxLayoutMain;
@@ -29,6 +30,25 @@ MainWindow::MainWindow(QWidget *parent)
     connect(udp, SIGNAL(readyRead()), this , SLOT(packet_arrived()));
     ui->tableWidget->setHorizontalHeaderLabels(QStringList() << "" << "ИД радиостанции" << "Владелец");
     ui->widgetControl->setVisible(false);
+
+    file.setFileName("settings.txt");
+    if(file.open(QIODevice::ReadOnly))
+    {
+        QByteArray ba;
+        while(!file.atEnd())//заносим из файла все радио в вектор
+        {
+            ba = file.readLine();
+            QList<QByteArray> lba = ba.split(',');
+            radio = new Radio(lba[0].toInt(), QString(lba[1]).trimmed());
+            radio->m_regNum = lba[2].trimmed().toInt();
+            v_rad.push_back(radio);
+
+        }
+    }
+    else
+        qDebug() << "not file!!!";
+    qDebug() << v_rad.size();
+
 }
 
 MainWindow::~MainWindow()
@@ -42,14 +62,29 @@ void MainWindow::on_pushButton_clicked()
     //rad->flush_registration();
     //for(int i = 0; i < v_rad.size(); i++)
     //    v_rad.at(i)->setWidth();
-   // ui->widget->setVisible(false);
+    // ui->widget->setVisible(false);
 
-//    char arr[] = {0,0,0,1,1};
-//    QByteArray data = QByteArray::fromRawData(arr, sizeof(arr));
-//    udp->writeDatagram(data, QHostAddress("127.0.0.1"), 4004);
-//    ui->widgetControl->setVisible(false);
+    //    char arr[] = {0,0,0,1,1};
+    //    QByteArray data = QByteArray::fromRawData(arr, sizeof(arr));
+    //    udp->writeDatagram(data, QHostAddress("127.0.0.1"), 4004);
+    //    ui->widgetControl->setVisible(false);
 
     ui->widgetControl->setVisible(true);
+    quint32 rc = 1;// начальная строка
+    ui->tableWidget->setRowCount(v_rad.size()+1);// сколько всего строк будет видно
+    foreach(Radio* r, v_rad)
+    {
+        QTableWidgetItem *twi = new QTableWidgetItem(tr("%1").arg(rc));
+        ui->tableWidget->setItem(rc,0, twi);
+
+        QTableWidgetItem *twi1 = new QTableWidgetItem(QString::number(r->m_radioNum));
+        ui->tableWidget->setItem(rc,1, twi1);
+
+        QTableWidgetItem *twi2 = new QTableWidgetItem(r->m_name);
+        ui->tableWidget->setItem(rc,2, twi2);
+
+        rc++;
+    }
 }
 
 void MainWindow::packet_arrived()
@@ -75,47 +110,54 @@ void MainWindow::on_pbAdd_clicked()
 void MainWindow::on_pbControlExit_clicked()
 {
     ui->widgetControl->setVisible(false);
+    // заполняем таблицу данными из вектора
+
+
 }
 
 void MainWindow::on_pbControlAdd_clicked()
 {
     // check!!!
+    bool check_number = true;
     quint32 num = ui->leControlID->text().toInt();
     foreach(Radio* r, v_rad)
         if (r->m_radioNum == num )
         {
-            QMessageBox::information();
+            check_number = false;
+            QMessageBox::information(this,"Информация","Такой ИД уже есть");
         }
-    radio = new Radio(num, ui->leControlName->text());
-    radCnt++;
-    v_rad.push_back(radio);
+    if(check_number)
+    {
+        radio = new Radio(num, ui->leControlName->text());
+        radio->m_regNum = 0;
+        v_rad.push_back(radio);
 
-    quint32 rc = ui->tableWidget->rowCount();
-    ui->tableWidget->setRowCount(rc+1);
+        quint32 rc = ui->tableWidget->rowCount();
+        ui->tableWidget->setRowCount(rc+1);
 
-    QTableWidgetItem *twi = new QTableWidgetItem(tr("%1").arg(rc+1));
-    ui->tableWidget->setItem(rc,0, twi);
-    QTableWidgetItem *twi1 = new QTableWidgetItem(ui->leControlID->text());
-    ui->tableWidget->setItem(rc,1, twi1);
-    QTableWidgetItem *twi2 = new QTableWidgetItem(ui->leControlName->text());
-    ui->tableWidget->setItem(rc,2, twi2);
-
+        QTableWidgetItem *twi = new QTableWidgetItem(tr("%1").arg(rc+1));
+        ui->tableWidget->setItem(rc,0, twi);
+        QTableWidgetItem *twi1 = new QTableWidgetItem(ui->leControlID->text());
+        ui->tableWidget->setItem(rc,1, twi1);
+        QTableWidgetItem *twi2 = new QTableWidgetItem(ui->leControlName->text());
+        ui->tableWidget->setItem(rc,2, twi2);
+    }
 
 }
 void MainWindow::on_pbControlDel_clicked()
 {
     quint32 radId = (ui->tableWidget->item(ui->tableWidget->currentRow(),1))->text().toInt();
-    QTableWidgetItem *twi0 = ui->tableWidget->item(ui->tableWidget->currentRow(),0);
-    QTableWidgetItem *twi1 = ui->tableWidget->item(ui->tableWidget->currentRow(),1);
-    QTableWidgetItem *twi2 = ui->tableWidget->item(ui->tableWidget->currentRow(),2);
+    //    QTableWidgetItem *twi0 = ui->tableWidget->item(ui->tableWidget->currentRow(),0);
+    //    QTableWidgetItem *twi1 = ui->tableWidget->item(ui->tableWidget->currentRow(),1);
+    //    QTableWidgetItem *twi2 = ui->tableWidget->item(ui->tableWidget->currentRow(),2);
     ui->tableWidget->removeRow(ui->tableWidget->currentRow());
-    delete twi0;
-    delete twi1;
-    delete twi2;
+    //    delete twi0;
+    //    delete twi1;
+    //    delete twi2;
     // стираем это радио из вектора
     qDebug() << v_rad.size();
     foreach(Radio* r, v_rad)
         if (r->m_radioNum == radId )
             v_rad.removeOne(r);
-    qDebug() << v_rad.size();
+    // qDebug() << v_rad.size();
 }
